0309 수정1

+++ server_member/src/main/java/org/zerock/server_member/service/ItemService.java
@@ -22,17 +22,25 @@ public class ItemService {
         return  itemRepository.save(item);
     }
     //삭제
+    @Transactional
     public void delete(Long id){
+        if (!itemRepository.existsById(id)) {
+            throw new IllegalArgumentException("해당 ID의 아이템이 존재하지 않습니다: " + id);
+        }
         itemRepository.deleteById(id);
     }
+
     //수정
     @Transactional
     public Item update(Long id, Item member){
-        Item i = itemRepository.findById(id).get();
+        Item i = itemRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 ID의 아이템이 존재하지 않습니다: " + id));
+
         i.setName(member.getName());   //member <- 수정 정보 객체
         i.setCategory(member.getCategory());
         i.setReleaseDate(member.getReleaseDate());
         i.setExpirationDate(member.getExpirationDate());
         return i;
     }
+
 }
